requirements.txt
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.1+cu113
torchvision==0.11.2+cu113
easydict==1.10
imageio==2.16.2
ipdb==0.13.11
lpips==0.1.4
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.4.0
PyYAML==6.0
scikit_image==0.19.2
sk-video==1.1.10
termcolor==2.1.1
tqdm==4.64.0
tensorboard==2.10.0
setuptools==59.5.0
